+++ util/time/microseconds.go
@@ -2,6 +2,6 @@ package time
 
 import "time"
 
-func MicroSeconds(t time.Time) uint64 {
-	return uint64(t.UnixNano() / 1000)
+func MicroSeconds(t time.Time) int64 {
+	return t.UnixNano() / 1000
 }
